+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Revision history for hmock
 
 ## 0.x.0.0 -- 2021-??-??
+* HMock now depends on `Predicate` from the `explainable-predicates` package.
 
 ## 0.4.0.0 -- 2021-08-22